openacc: Implement fix for erroneous task completion polling.
It is possible that getFinishedTasks is called (at random times) before
the actual task kernel launch. This will result in erroneous detection
of the task having finished, as acc_async_test() will return true (as,
indeed, there is nothing running on the queue).

To work around that, we introduced a _launched flag in the OpenAccQueue
which is set to 0 on initialization; it is then set to 1 on PostBody
to signal that now the queue can be tested for completion.