openacc: Introduced device index in queues and environment; doing set_device on finished()
It appears that acc_async_test() call is only guaranteed to work correctly for
asynchronous regions launched by the *same host thread*. Obviously, this is
a problem for OmpSs, but we found that calling acc_set_device seems to help
the current thread also detect asyncs launched by other threads.

Thus, since device index is becoming important, I added it as a field in the
OpenAccQueue objects and device environment to be more easily accessed.