Changed devices initialization order.
Prioritized OpenACC vs CUDA. The reason for that is the way NVIDIA/PGI
runtime for OpenACC handles the underlying CUDA; if an OpenACC runtime
call (e.g. to get device num) is run after a CUDA runtime call it results
in broken CUDA context and the runtime is never initialized. Fortunately, the
other way around (CUDA call following OpenACC call) seems to work, so we
work-around it by changing the device detection and initialization
sequence in HardwareInfo.