openacc: extended openacc_LDFLAGS, solves plain mcxx programs not finding symbols issue
Previously, we were linking the runtime with the appropriate PGI libraries that provide the
symbols used in the runtime. That works fine when compiling with pgimcxx, that inherits
the automatic linking procedure of the PGI compilers. However, with (non-openacc) OmpSs-2
programs using mcxx lacking these flags, the loader would fail to load and start the program,
due to unresolved OpenACC related PGI internal symbols.

Providing extra dependent PGI libs in the runtime itself appears to solve the problem.